#51 fixed time zone
fixed time zone setting function and adding a spot in the setup to call the function. Tested by reinstalling linux on my laptop. Had some issues with github for windows wanting to auto change all 'end of line' characters to CRLF in stead of the LF linux apparently uses. Had to add a .gitattributes file to fix it, couldn't find any options in the app to adjust it >.>